separated handler into its own file to avoid verbosity; added notes on name/dob being required
